Add dark mode to menu bar for small screens
